device-gateway/requirements.txt
Flask
paho-mqtt
python-dotenv
APScheduler
requests
